requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
httpx
openai>=1.54.0
pydantic>=2.6.0
